refactor(app): move createAnnotationMemberCompletions out of the base DSL filter field
The annotation member expansion is domain vocabulary shared by the two
DSL consumers, not part of the generic field. The base component now
exposes createLoadedCompletionSection as the extension point for loaded
completion groups.
