+++ app/src/components/filter/index.tsx
@@ -1,2 +1,3 @@
 export * from "./Toolbar";
+export * from "./annotationCompletions";
 export * from "./DSLFilterConditionField";
